+++ src/Auditable.php
@@ -7,17 +7,18 @@ use Illuminate\Database\Eloquent\SoftDeletes;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Facades\App;
 use Illuminate\Support\Facades\Config;
+use Illuminate\Support\Facades\Event;
 use OwenIt\Auditing\Contracts\AttributeEncoder;
 use OwenIt\Auditing\Contracts\AttributeRedactor;
-use OwenIt\Auditing\Contracts\IpAddressResolver;
-use OwenIt\Auditing\Contracts\UrlResolver;
-use OwenIt\Auditing\Contracts\UserAgentResolver;
-use OwenIt\Auditing\Contracts\UserResolver;
+use OwenIt\Auditing\Contracts\Resolver;
+use OwenIt\Auditing\Events\AuditCustom;
 use OwenIt\Auditing\Exceptions\AuditableTransitionException;
 use OwenIt\Auditing\Exceptions\AuditingException;
 
 trait Auditable
 {
+
+
     /**
      * Auditable attributes excluded from the Audit.
      *
@@ -38,6 +39,24 @@ trait Auditable
      * @var bool
      */
     public static $auditingDisabled = false;
+
+    /**
+     * Property may set custom event data to register
+     * @var null|array
+     */
+    public $auditCustomOld = null;
+
+    /**
+     * Property may set custom event data to register
+     * @var null|array
+     */
+    public $auditCustomNew = null;
+
+    /**
+     * If this is a custom event (as opposed to an eloquent event
+     * @var bool
+     */
+    public $isCustomEvent = false;
 
     /**
      * Auditable boot logic.
@@ -105,6 +124,22 @@ trait Auditable
     }
 
     /**
+     * @return array
+     */
+    public function getAuditExclude(): array
+    {
+        return $this->auditExclude ?? Config::get('audit.exclude', []);
+    }
+
+    /**
+     * @return array
+     */
+    public function getAuditInclude(): array
+    {
+        return $this->auditInclude ?? [];
+    }
+
+    /**
      * Get the old/new attributes of a retrieved event.
      *
      * @return array
@@ -138,6 +173,14 @@ trait Auditable
         return [
             [],
             $new,
+        ];
+    }
+
+    protected function getCustomEventAttributes(): array
+    {
+        return [
+            $this->auditCustomOld,
+            $this->auditCustomNew
         ];
     }
 
@@ -205,6 +248,10 @@ trait Auditable
             return false;
         }
 
+        if ($this->isCustomEvent) {
+            return true;
+        }
+
         return $this->isEventAuditable($this->auditEvent);
     }
 
@@ -262,7 +309,7 @@ trait Auditable
 
         list($old, $new) = $this->$attributeGetter();
 
-        if ($this->getAttributeModifiers()) {
+        if ($this->getAttributeModifiers() && !$this->isCustomEvent) {
             foreach ($old as $attribute => $value) {
                 $old[$attribute] = $this->modifyAttributeValue($attribute, $value);
             }
@@ -278,7 +325,7 @@ trait Auditable
 
         $user = $this->resolveUser();
 
-        return $this->transformAudit([
+        return $this->transformAudit(array_merge([
             'old_values'           => $old,
             'new_values'           => $new,
             'event'                => $this->auditEvent,
@@ -286,11 +333,8 @@ trait Auditable
             'auditable_type'       => $this->getMorphClass(),
             $morphPrefix . '_id'   => $user ? $user->getAuthIdentifier() : null,
             $morphPrefix . '_type' => $user ? $user->getMorphClass() : null,
-            'url'                  => $this->resolveUrl(),
-            'ip_address'           => $this->resolveIpAddress(),
-            'user_agent'           => $this->resolveUserAgent(),
             'tags'                 => empty($tags) ? null : $tags,
-        ]);
+        ], $this->runResolvers()));
     }
 
     /**
@@ -310,67 +354,29 @@ trait Auditable
      */
     protected function resolveUser()
     {
-        $userResolver = Config::get('audit.resolver.user');
+        $userResolver = Config::get('audit.user.resolver');
 
-        if (is_subclass_of($userResolver, UserResolver::class)) {
+        if (is_subclass_of($userResolver, \OwenIt\Auditing\Contracts\UserResolver::class)) {
             return call_user_func([$userResolver, 'resolve']);
         }
 
         throw new AuditingException('Invalid UserResolver implementation');
     }
 
-    /**
-     * Resolve the URL.
-     *
-     * @return string
-     * @throws AuditingException
-     *
-     */
-    protected function resolveUrl(): string
+    protected function runResolvers(): array
     {
-        $urlResolver = Config::get('audit.resolver.url');
+        $resolved = [];
+        foreach (Config::get('audit.resolvers', []) as $name => $implementation) {
+            if (empty($implementation)) {
+                continue;
+            }
 
-        if (is_subclass_of($urlResolver, UrlResolver::class)) {
-            return call_user_func([$urlResolver, 'resolve']);
+            if (!is_subclass_of($implementation, Resolver::class)) {
+                throw new AuditingException('Invalid Resolver implementation for: ' . $name);
+            }
+            $resolved[$name] = call_user_func([$implementation, 'resolve'], $this);
         }
-
-        throw new AuditingException('Invalid UrlResolver implementation');
-    }
-
-    /**
-     * Resolve the IP Address.
-     *
-     * @return string
-     * @throws AuditingException
-     *
-     */
-    protected function resolveIpAddress(): string
-    {
-        $ipAddressResolver = Config::get('audit.resolver.ip_address');
-
-        if (is_subclass_of($ipAddressResolver, IpAddressResolver::class)) {
-            return call_user_func([$ipAddressResolver, 'resolve']);
-        }
-
-        throw new AuditingException('Invalid IpAddressResolver implementation');
-    }
-
-    /**
-     * Resolve the User Agent.
-     *
-     * @return string|null
-     * @throws AuditingException
-     *
-     */
-    protected function resolveUserAgent()
-    {
-        $userAgentResolver = Config::get('audit.resolver.user_agent');
-
-        if (is_subclass_of($userAgentResolver, UserAgentResolver::class)) {
-            return call_user_func([$userAgentResolver, 'resolve']);
-        }
-
-        throw new AuditingException('Invalid UserAgentResolver implementation');
+        return $resolved;
     }
 
     /**
@@ -417,6 +423,10 @@ trait Auditable
     {
         if (empty($event)) {
             return;
+        }
+
+        if ($this->isCustomEvent) {
+            return 'getCustomEventAttributes';
         }
 
         foreach ($this->getAuditEvents() as $key => $value) {
@@ -493,22 +503,6 @@ trait Auditable
         }
 
         return Config::get('audit.enabled', true);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getAuditInclude(): array
-    {
-        return $this->auditInclude ?? [];
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getAuditExclude(): array
-    {
-        return $this->auditExclude ?? [];
     }
 
     /**
@@ -611,5 +605,105 @@ trait Auditable
         }
 
         return $this;
+    }
+
+    /*
+    |--------------------------------------------------------------------------
+    | Pivot help methods
+    |--------------------------------------------------------------------------
+    |
+    | Methods for auditing pivot actions
+    |
+    */
+
+    /**
+     * @param string $relationName
+     * @param mixed $id
+     * @param array $attributes
+     * @param bool $touch
+     * @return void
+     * @throws AuditingException
+     */
+    public function auditAttach(string $relationName, $id, array $attributes = [], $touch = true, $columns = ['name'])
+    {
+        if (!method_exists($this, $relationName) || !method_exists($this->{$relationName}(), 'attach')) {
+            throw new AuditingException('Relationship ' . $relationName . ' was not found or does not support method attach');
+        }
+        $this->auditEvent = 'attach';
+        $this->isCustomEvent = true;
+        $this->auditCustomOld = [
+            $relationName => $this->{$relationName}()->get()->isEmpty() ? [] : $this->{$relationName}()->get()->toArray()
+        ];
+        $this->{$relationName}()->attach($id, $attributes, $touch);
+        $this->auditCustomNew = [
+            $relationName => $this->{$relationName}()->get()->isEmpty() ? [] : $this->{$relationName}()->get()->toArray()
+        ];
+        Event::dispatch(AuditCustom::class, [$this]);
+    }
+
+    /**
+     * @param string $relationName
+     * @param mixed $ids
+     * @param bool $touch
+     * @return int
+     * @throws AuditingException
+     */
+    public function auditDetach(string $relationName, $ids = null, $touch = true)
+    {
+        if (!method_exists($this, $relationName) || !method_exists($this->{$relationName}(), 'detach')) {
+            throw new AuditingException('Relationship ' . $relationName . ' was not found or does not support method detach');
+        }
+
+        $this->auditEvent = 'detach';
+        $this->isCustomEvent = true;
+        $this->auditCustomOld = [
+            $relationName => $this->{$relationName}()->get()->isEmpty() ? [] : $this->{$relationName}()->get()->toArray()
+        ];
+        $results = $this->{$relationName}()->detach($ids, $touch);
+        $this->auditCustomNew = [
+            $relationName => $this->{$relationName}()->get()->isEmpty() ? [] : $this->{$relationName}()->get()->toArray()
+        ];
+        Event::dispatch(AuditCustom::class, [$this]);
+        return empty($results) ? 0 : $results;
+    }
+
+    /**
+     * @param $relationName
+     * @param \Illuminate\Support\Collection|\Illuminate\Database\Eloquent\Model|array $ids
+     * @param bool $detaching
+     * @return array
+     * @throws AuditingException
+     */
+    public function auditSync($relationName, $ids, $detaching = true)
+    {
+        if (!method_exists($this, $relationName) || !method_exists($this->{$relationName}(), 'sync')) {
+            throw new AuditingException('Relationship ' . $relationName . ' was not found or does not support method sync');
+        }
+
+        $this->auditEvent = 'sync';
+        $this->isCustomEvent = true;
+        $this->auditCustomOld = [
+            $relationName => $this->{$relationName}()->get()->isEmpty() ? [] : $this->{$relationName}()->get()->toArray()
+        ];
+        $changes = $this->{$relationName}()->sync($ids, $detaching);
+        $this->auditCustomNew = [
+            $relationName => $this->{$relationName}()->get()->isEmpty() ? [] : $this->{$relationName}()->get()->toArray()
+        ];
+        Event::dispatch(AuditCustom::class, [$this]);
+        return $changes;
+    }
+
+    /**
+     * @param string $relationName
+     * @param \Illuminate\Support\Collection|\Illuminate\Database\Eloquent\Model|array $ids
+     * @return array
+     * @throws AuditingException
+     */
+    public function auditSyncWithoutDetaching(string $relationName, $ids)
+    {
+        if (!method_exists($this, $relationName) || !method_exists($this->{$relationName}(), 'syncWithoutDetaching')) {
+            throw new AuditingException('Relationship ' . $relationName . ' was not found or does not support method syncWithoutDetaching');
+        }
+        return $this->auditSync($relationName, $ids, false);
     }
 }