+++ src/Auditor.php
@@ -2,6 +2,7 @@
 
 namespace OwenIt\Auditing;
 
+use Illuminate\Support\Facades\Config;
 use Illuminate\Support\Manager;
 use InvalidArgumentException;
 use OwenIt\Auditing\Contracts\Auditable;
@@ -66,9 +67,23 @@ class Auditor extends Manager implements Contracts\Auditor
             return;
         }
 
-        if ($audit = $driver->audit($model)) {
-            $driver->prune($model);
+        // If we want to avoid storing Audits with empty old_values & new_values, return null here.
+        if (!Config::get('audit.empty_values')) {
+            if (
+                empty($model->toAudit()['new_values']) &&
+                empty($model->toAudit()['old_values']) &&
+                !in_array($model->getAuditEvent(), Config::get('audit.allowed_empty_values'))
+            ) {
+                return;
+            }
         }
+
+        $audit = $driver->audit($model);
+        if (!$audit) {
+            return;
+        }
+
+        $driver->prune($model);
 
         $this->container->make('events')->dispatch(
             new Audited($model, $driver, $audit)
@@ -88,15 +103,16 @@ class Auditor extends Manager implements Contracts\Auditor
     /**
      * Fire the Auditing event.
      *
-     * @param \OwenIt\Auditing\Contracts\Auditable   $model
+     * @param \OwenIt\Auditing\Contracts\Auditable $model
      * @param \OwenIt\Auditing\Contracts\AuditDriver $driver
      *
      * @return bool
      */
     protected function fireAuditingEvent(Auditable $model, AuditDriver $driver): bool
     {
-        return $this->container->make('events')->until(
-            new Auditing($model, $driver)
-        ) !== false;
+        return $this
+                ->container
+                ->make('events')
+                ->until(new Auditing($model, $driver)) !== false;
     }
 }