+++ src/Console/AuditDriverCommand.php
@@ -26,7 +26,7 @@ class AuditDriverCommand extends GeneratorCommand
      */
     protected function getStub()
     {
-        return __DIR__ . '/../../drivers/driver.stub';
+        return __DIR__ . '/../../stubs/driver.stub';
     }
 
     /**