+++ src/Contracts/Resolver.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace OwenIt\Auditing\Contracts;
+
+interface Resolver
+{
+    public static function resolve(Auditable $auditable);
+}