+++ src/Resolvers/IpAddressResolver.php
@@ -3,13 +3,12 @@
 namespace OwenIt\Auditing\Resolvers;
 
 use Illuminate\Support\Facades\Request;
+use OwenIt\Auditing\Contracts\Auditable;
+use OwenIt\Auditing\Contracts\Resolver;
 
-class IpAddressResolver implements \OwenIt\Auditing\Contracts\IpAddressResolver
+class IpAddressResolver implements Resolver
 {
-    /**
-     * {@inheritdoc}
-     */
-    public static function resolve(): string
+    public static function resolve(Auditable $auditable): string
     {
         return Request::ip();
     }