+++ src/Resolvers/UrlResolver.php
@@ -4,13 +4,14 @@ namespace OwenIt\Auditing\Resolvers;
 
 use Illuminate\Support\Facades\App;
 use Illuminate\Support\Facades\Request;
+use OwenIt\Auditing\Contracts\Auditable;
 
-class UrlResolver implements \OwenIt\Auditing\Contracts\UrlResolver
+class UrlResolver implements \OwenIt\Auditing\Contracts\Resolver
 {
     /**
-     * {@inheritdoc}
+     * @return string
      */
-    public static function resolve(): string
+    public static function resolve(Auditable $auditable): string
     {
         if (App::runningInConsole()) {
             return 'console';