+++ tests/Models/Category.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace OwenIt\Auditing\Tests\Models;
+
+class Category extends \Illuminate\Database\Eloquent\Model
+{
+}